dto and model rename
