+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/User/GetUserAuditLogResponse.cs
@@ -18,4 +18,4 @@ namespace Energinet.DataHub.MarketParticipant.Application.Commands.User;
 
 public sealed record GetUserAuditLogResponse(
     IEnumerable<UserRoleAssignmentAuditLogEntryDto> UserRoleAssignmentAuditLogs,
-    IEnumerable<UserInviteDetailsAuditLogDto> InviteDetailsAuditLogs);
+    IEnumerable<UserInviteDetailsAuditLogEntryDto> InviteDetailsAuditLogs);
