+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/User/UserInviteDetailsAuditLogEntryDto.cs
@@ -11,13 +11,14 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 using System;
 
-namespace Energinet.DataHub.MarketParticipant.Domain.Model.Users;
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.User;
 
-public sealed record UserInviteDetailsAuditLog(
-    UserId UserId,
-    UserId ChangedByUserId,
+public sealed record UserInviteDetailsAuditLogEntryDto(
+    Guid UserId,
+    Guid ChangedByUserId,
     Guid ActorId,
     string ActorName,
     DateTimeOffset Timestamp);
