+++ source/Energinet.DataHub.MarketParticipant.Domain/Services/IUserInvitationService.cs
@@ -26,6 +26,6 @@ public interface IUserInvitationService
     /// Invites the user specified in the invitation into DataHub.
     /// </summary>
     /// <param name="invitation">An invitation of a new or existing user.</param>
-    /// <param name="invitationSentByUserId">user sending til invitation</param>
+    /// <param name="invitationSentByUserId">user sending the invitation</param>
     Task InviteUserAsync(UserInvitation invitation, UserId invitationSentByUserId);
 }
